pyproject.toml
[project]
name = "smartdoc"
version = "0.4.0"
description = "A quality scorer for OAS files"

dependencies = [
  "typer>=0.9",
  "rich",
  "pyyaml",
  "openai",
  "python-dotenv",
  "httpx",
  "pydantic",
]

[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project.scripts]
smartdoc = "smartdoc.cli:app"
